readme updated

+++ src/include/string.h
@@ -5,6 +5,6 @@
 
 uint32_t strlen(const char* s);
 
-char* itoa(uint32_t val);
+char* itoa(uint64_t val);
 
 #endif
